Send Btc fix amount parse

+++ CCP/android/reference_wallet/fermat-ccp-android-reference-wallet-bitcoin-wallet-bitdubai/src/main/java/com/bitdubai/reference_niche_wallet/bitcoin_wallet/fragments/wallet_final_version/SendFormFragment.java
@@ -501,7 +501,7 @@ public class SendFormFragment extends AbstractFermatFragment<ReferenceWalletSess
                     EditText txtAmount = (EditText) rootView.findViewById(R.id.amount);
                     String amount = txtAmount.getText().toString();
 
-                    if(!amount.equals("") && amount!=null && Long.parseLong(amount)!=0) {
+                    if(!amount.equals("") && amount!=null && Double.parseDouble(amount) != 0) {
                         try {
                             String notes=null;
                             if(txt_notes.getText().toString().length()!=0){
